requirements.txt
mysql-connector-python==9.4.0
python-dotenv==1.2.1
Faker==19.13.0
